Re-structure Slack connector to handle Webhook in the future
